requirements.txt
# Ergane — Job Search Automation Tool
# Python 3.11+ compatible

# Scraping
scrapling>=0.4.3
playwright==1.40.0
beautifulsoup4==4.12.3
requests==2.31.0

# Scheduling
APScheduler==3.10.4

# Telegram Bot
python-telegram-bot==21.0

# Utilities
python-dotenv==1.0.0
pyyaml>=6.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Document Generation
python-docx==1.1.0

# LangChain + LangGraph for AI agents
langchain>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
pydantic>=2.0
